+++ CHANGELOG.md
@@ -114,6 +114,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `lt_r_mem_data_read()`: Renamed parameter `size` to `read_size`.
 - Added new CMake option `LT_CRYPTO` for selecing crypto provider with a string, renamed `LT_USE_TREZOR_CRYPTO` to `LT_CRYPTO_TREZOR`.
 - Reworked the documentation using the MkDocs framework.
+- Renamed `lt_session_state_t` to `lt_host_eph_keys_t`.
+- Renamed `lt_secure_session_state_t` to `lt_secure_session_status_t`.
+- Renamed `lt_l3_state_t.session` to `lt_l3_state_t.session_status`.
 
 ### Added
 - CMake option for setting logging verbosity level: `LT_LOG_LVL`.